changed project structure for easier readability of proj workflow and cleaner build.

+++ src/boot/VGA_driver.c
@@ -1,4 +1,5 @@
 #include "VGA_driver.h"
+#include <stddef.h>
 
 #define FRAMEBUFFER_ADDRESS_OFFSET 8
 #define FRAMEBUFFER_HEIGHT_OFFSET 24
@@ -9,8 +10,8 @@ struct Framebuffer framebuffer;
 
 void print(void){
     
-    for(int x = 0; x < framebuffer.width; x++){
-        for(int y=0; y< framebuffer.height; y++){
+    for(size_t x = 0; x < framebuffer.width; x++){
+        for(size_t y=0; y< framebuffer.height; y++){
             *(framebuffer.starting_address + (y*framebuffer.width + x)*3 + framebuffer.pixel.red_bits_index/8) = 0xFF;
             }
         }
